+++ src/main/java/com/sparta/ticketauction/domain/admin/response/GradeCreateResponse.java
@@ -3,12 +3,12 @@ package com.sparta.ticketauction.domain.admin.response;
 import lombok.Getter;
 
 @Getter
-public class GradeResponse {
+public class GradeCreateResponse {
 	private final Long placeId;
 
 	private final Long gradeId;
 
-	public GradeResponse(Long placeId, Long gradeId) {
+	public GradeCreateResponse(Long placeId, Long gradeId) {
 		this.placeId = placeId;
 		this.gradeId = gradeId;
 	}
